add TS functionality
