+++ .gitignore
@@ -3,3 +3,7 @@ node_modules
 
 # CatchAll for Accidental GeoJSON Output Testing
 *.geojson
+
+
+# Remove build files
+/build
