+++ src/utils/request.ts
@@ -1,6 +1,6 @@
 // src/utils/request.mjs
 
-import { logging } from '../utils/logging.mjs'
+import { logging } from './logging.js'
 
 import fetch from 'node-fetch' // not required in Node17.5 (LTS) onwards
