+++ tsconfig.json
@@ -1,15 +1,10 @@
 {
-  "compilerOptions": {
-    "target": "es2016",
-    "lib": ["es6"],
-    "module": "commonjs",
-    "rootDir": "src",
-    "resolveJsonModule": true,
-    "allowJs": true,
-    "outDir": "build",
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "strict": true,
-    "noImplicitAny": true,
+    "extends": "@tsconfig/node16/tsconfig.json",
+    "compilerOptions": {
+        "outDir": "build",
+        "module": "es6",
+        "target": "es6",
+    },
+    "include": ["src"],
+    "exclude": ["node_modules"]
   }
-}
